requirements.txt
smolagents
litellm
python-dotenv
requests
browser-use
langchain-openai 